The "Release" folder in the build directory is now checked for header files during Ad Hoc and App Store builds.
Most static libararies have a Debug and Release build configuration.
Since these project templates do not have a Release build configuration
the header files for static libraries that have been added as a
dependency are not able to be found during Ad Hoc and App Store builds.

The header and user header search paths have been modified so that they
will look for a "Release" folder in the build directory which is
typically what is used for Ad Hoc and App Store builds.