Replaced TestFlight with HockeyApp.
I am still not 100% convinced that this is the “correct” way to do this.

Everything looks almost identical to if you dragged and dropped the
embedded framework into the project but I had to copy the embedded
framework in pieces to get that to work. Also the “Resources” folder no
longer has symbolic links to the resources inside the framework.
Xcode’s project templating system seems to deep copy the resources
instead of copying the symbolic link. This honestly shouldn’t be an
issue because the resources inside the framework won’t change unless
the entire framework is updated.